requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.2.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
